requirements.txt
allure-pytest==2.13.5
pytest==8.3.3
pytest-xdist==3.6.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.24.0